requirements.txt
# Django 프로젝트 필수 패키지 목록 - DN_SOLUTION2 리모델링
# 1,000 RPS 처리 최적화

# =============================================================================
# 핵심 프레임워크 (필수)
# =============================================================================
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# =============================================================================
# 데이터베이스 (운영 환경)
# =============================================================================
psycopg2-binary==2.9.10       # PostgreSQL 드라이버
django-model-utils==4.3.1      # 모델 유틸리티

# =============================================================================
# 캐싱 및 성능 (필수)
# =============================================================================
redis==5.0.1                   # Redis 클라이언트
django-redis==5.4.0           # Django Redis 백엔드
django-cachalot==2.6.1        # ORM 캐시
django-cache-machine==1.2.0   # 자동 캐시 무효화

# =============================================================================
# 비동기 작업 처리 (필수)
# =============================================================================
celery==5.3.4                 # 비동기 작업 큐
django-celery-beat==2.5.0     # Celery 스케줄러
django-celery-results==2.5.1  # Celery 결과 저장
kombu==5.3.4                  # 메시지 브로커

# =============================================================================
# WebSocket 지원 (실시간 알림)
# =============================================================================
channels==4.0.0               # WebSocket 지원
channels-redis==4.1.0         # Redis 채널 레이어
daphne==4.0.0                 # ASGI 서버

# =============================================================================
# 보안 및 인증 (필수)
# =============================================================================
django-cors-headers==4.3.1    # CORS 처리
django-ratelimit==4.1.0       # API 속도 제한
django-axes==6.1.1            # 로그인 보호
django-oauth-toolkit==1.7.1   # OAuth2 지원 (선택)

# =============================================================================
# 환경 변수 및 설정 관리 (필수)
# =============================================================================
python-decouple==3.8
python-dotenv==1.0.0
django-environ==0.11.2

# =============================================================================
# API 및 데이터 처리 (필수)
# =============================================================================
django-filter==23.4           # 필터링
Pillow==10.1.0                # 이미지 처리
requests==2.31.0              # HTTP 클라이언트
openpyxl==3.1.2              # Excel 파일 처리
pandas==2.1.4                 # 데이터 분석

# =============================================================================
# 로깅 및 모니터링 (운영 환경)
# =============================================================================
colorlog==6.8.0               # 컬러 로깅
structlog==23.2.0             # 구조화된 로깅
django-prometheus==2.3.1      # Prometheus 메트릭

# =============================================================================
# API 문서화 (개발/운영)
# =============================================================================
drf-spectacular==0.27.0       # OpenAPI 3.0 문서화

# =============================================================================
# 테스팅 도구 (개발 환경)
# =============================================================================
pytest==7.4.3                 # 테스트 프레임워크
pytest-django==4.7.0          # Django 테스트 지원
pytest-cov==4.1.0            # 커버리지 측정
factory-boy==3.3.0            # 테스트 데이터 생성
faker==20.1.0                 # 가짜 데이터 생성

# =============================================================================
# 배포 및 성능 최적화 (운영 환경)
# =============================================================================
gunicorn==21.2.0              # WSGI 서버
gevent==23.9.1                # 비동기 워커
whitenoise==6.6.0             # 정적 파일 서빙
django-storages==1.14.2       # 클라우드 스토리지
boto3==1.35.7                 # AWS SDK

# =============================================================================
# 에러 모니터링 (운영 환경)
# =============================================================================
sentry-sdk[django]==1.40.4    # 에러 추적

# =============================================================================
# 이메일 (운영 환경)
# =============================================================================
django-anymail==10.2          # 이메일 서비스 통합

# =============================================================================
# 데이터 검증 및 시리얼라이제이션
# =============================================================================
marshmallow==3.20.2          # 데이터 스키마
pydantic==2.5.2              # 데이터 검증

# =============================================================================
# 유틸리티
# =============================================================================
pytz==2023.3                 # 시간대
python-dateutil==2.8.2       # 날짜 처리
uuid==1.30                   # UUID 처리
shortuuid==1.0.11            # 짧은 UUID

# =============================================================================
# HTTP 및 네트워킹
# =============================================================================
httpx==0.25.2                # 비동기 HTTP 클라이언트
urllib3==2.1.0               # HTTP 라이브러리